requirements.test.txt
# Strictly for tests
coverage==7.2.1
pytest==7.2.2
pytest-asyncio==0.20.3
pytest-cov==3.0.0
pytest-homeassistant-custom-component==0.13.20